+++ src/class-supported-post-types.php
@@ -32,7 +32,7 @@ final class Supported_Post_Types {
 		// Get all post types.
 		$post_types                 = get_post_types( [], 'objects' );
 		$supported_post_types       = array_filter( $post_types, fn( $type ) => $type->public && use_block_editor_for_post_type( $type->name ) );
-		$this->supported_post_types = ( wp_list_pluck( $supported_post_types, 'name' ) );
+		$this->supported_post_types = array_keys( wp_list_pluck( $supported_post_types, 'name' ) );
 		$this->register_post_meta();
 	}
 
@@ -42,6 +42,11 @@ final class Supported_Post_Types {
 	 * @return string[]
 	 */
 	public function get_supported_post_types(): array {
+		/**
+		 * Filter the WP Curate supported post types.
+		 *
+		 * @param string[] $supported_post_types The supported post types.
+		 */
 		return apply_filters( 'wp_curate_supported_post_types', $this->supported_post_types );
 	}
 
@@ -63,9 +68,13 @@ final class Supported_Post_Types {
 				$post_id   = absint( $_GET['post'] );
 				$post_type = get_post_type( $post_id );
 			// phpcs:ignore WordPress.VIP.SuperGlobalInputUsage.AccessDetected, WordPress.Security.NonceVerification.NoNonceVerification, WordPress.Security.NonceVerification.Recommended
-			} elseif ( 'post-new.php' === $pagenow && ! empty( $_GET['post_type'] ) ) {
-				// phpcs:ignore WordPress.VIP.SuperGlobalInputUsage.AccessDetected, WordPress.Security.NonceVerification.Recommended
-				$post_type = sanitize_text_field( $_GET['post_type'] );
+			} elseif ( 'post-new.php' === $pagenow ) {
+				if ( ! empty( $_GET['post_type'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+					$post_type = sanitize_text_field( $_GET['post_type'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+				} else {
+					// Default to post.
+					$post_type = 'post';
+				}
 			}
 		}
 		return $post_type;
@@ -83,5 +92,14 @@ final class Supported_Post_Types {
 				'type' => 'boolean',
 			]
 		);
+	}
+
+	/**
+	 * Should we register the block in the current context.
+	 *
+	 * @return boolean
+	 */
+	public function should_register_block(): bool {
+		return empty( $this->get_current_post_type() ) || in_array( $this->get_current_post_type(), $this->get_supported_post_types(), true );
 	}
 }